requirements.txt
streamlit==1.28.1
mysql-connector-python==8.2.0
pandas==2.1.1
plotly==5.17.0
numpy==1.24.3
python-dotenv==1.0.0
